+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [ refactor/cobros-total-calculo ] - 2025/12/07
+
+### Changed
+* `backend/service_pedidos/apps/cobros/views.py`
+  * Añade al viewset de cobros un nuevo método `total_by_date` que devuelve los ingresos brutos según los cobros que hubo en un día el cual se parsea a través de la url.
+* `Frontend/src/services/cobro_service.ts`
+  * Integra el nuevo endpoint de cobros al servicio del frontend.
+
 ## [ fix/front/recetas-styles ] - 2025/12/07
 
 ### Changed